+++ Assets/Scripts/Characters/Samurai.cs
@@ -6,14 +6,14 @@ namespace Characters
 {
     public class Samurai : Player
     {
-        private const float HP = 5;
-        private const float MP = 6;
-        private const float SPE = 10;
-        private const float RAN = 8;
-        private const float DMG = 10;
-        private const float MDMG = 1;
-        private const float DEF = 2;
-        private const float MDEF = 6;
+        private const int HP = 5;
+        private const int MP = 6;
+        private const int SPE = 10;
+        private const int RAN = 8;
+        private const int DMG = 10;
+        private const int MDMG = 1;
+        private const int DEF = 2;
+        private const int MDEF = 6;
         private const int LVL = 1;
         private void Awake()
         {
